Rm succinct (#24)
* Removed succinct/util
* Moved files
