+++ include/bitvector_collection.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "succinct/bit_vector.hpp"
+#include "bit_vector.hpp"
 
 #include "codec/compact_elias_fano.hpp"
