+++ include/util/intrinsics.hpp
